Add parser option for scraping cubdomain.com
