+++ main.py
@@ -40,6 +40,16 @@ if __name__ == "__main__":
         """,
     )
 
+    parser.add_argument(
+        "-c",
+        "--fetch-cubdomain",
+        dest="cubdomain",
+        action="store_true",
+        help="""
+        (EXPERIMENTAL) Download domains from cubdomain.com to .txt files.
+        """,
+    )
+
     group = parser.add_mutually_exclusive_group()
     group.add_argument(
         "-i",
@@ -112,14 +122,17 @@ if __name__ == "__main__":
 
     args = parser.parse_args()
     args.vendors = sorted([x.capitalize() for x in args.vendors])
-    if not (args.fetch or args.identify or args.retrieve):
+    if not (args.fetch or args.cubdomain or args.identify or args.retrieve):
         parser.error("No action requested, add -h for help")
 
     fetch: bool = args.fetch
+    cubdomain: bool = args.cubdomain
     identify: bool = args.identify
     use_existing_hashes: bool = args.use_existing_hashes
     retrieve: bool = args.retrieve
     sources: List[str] = args.sources
     vendors: List[str] = args.vendors
 
-    process_flags(fetch, identify, use_existing_hashes, retrieve, sources, vendors)
+    process_flags(
+        fetch, cubdomain, identify, use_existing_hashes, retrieve, sources, vendors
+    )
