+++ modules/scrape_cubdomain.py
@@ -4,73 +4,13 @@ from datetime import datetime, timedelta
 import os
 from collections import ChainMap
 from typing import Dict, List, Tuple
-import requests
-from requests.adapters import HTTPAdapter
-from urllib3.util.retry import Retry
 from bs4 import BeautifulSoup
-from logger_utils import init_logger
-from ray_utils import execute_with_ray
+from modules.logger_utils import init_logger
+from modules.ray_utils import execute_with_ray
+from modules.requests_utils import EnhancedSession
+
 
 logger = init_logger()
-
-DEFAULT_TIMEOUT = 10  # seconds
-
-
-class TimeoutHTTPAdapter(HTTPAdapter):
-    """HTTP Adapter with connection timeout
-
-    Args:
-        HTTPAdapter: The built-in HTTP Adapter for urllib3
-    """
-
-    def __init__(self, *args, **kwargs):
-        self.timeout = DEFAULT_TIMEOUT
-        if "timeout" in kwargs:
-            self.timeout = kwargs["timeout"]
-            del kwargs["timeout"]
-        super().__init__(*args, **kwargs)
-
-    def send(self, request, **kwargs):
-        # pylint: disable=arguments-differ
-        timeout = kwargs.get("timeout")
-        if timeout is None:
-            kwargs["timeout"] = self.timeout
-        return super().send(request, **kwargs)
-
-
-class EnhancedSession:
-    # pylint: disable=too-few-public-methods
-    """requests.Session() with connection timeout
-    + connection retries with backoff"""
-
-    def __init__(self):
-        retry_strategy = Retry(
-            total=3,
-            status_forcelist=[429, 500, 502, 503, 504],
-            allowed_methods=["GET"],
-            backoff_factor=1,
-        )
-        self.http = requests.Session()
-        assert_status_hook = (
-            lambda response, *args, **kwargs: response.raise_for_status()
-        )
-        self.http.hooks["response"] = [assert_status_hook]
-        self.http.headers.update(
-            {
-                "User-Agent": "Mozilla/5.0 (X11; Ubuntu; "
-                "Linux x86_64; rv:68.0) Gecko/20100101 Firefox/68.0"
-            }
-        )
-        self.http.mount("", TimeoutHTTPAdapter(max_retries=retry_strategy))
-
-    def get_session(self) -> requests.Session:
-        """getter
-
-        Returns:
-            requests.Session: requests.Session with
-            connection timeout + connection retries with backoff
-        """
-        return self.http
 
 
 def generate_dates_and_root_urls() -> Tuple[List[datetime], List[str]]:
@@ -93,24 +33,24 @@ def generate_dates_and_root_urls() -> Tuple[List[datetime], List[str]]:
     return dates, root_urls
 
 
-def create_root_url_map(http: requests.Session, date: datetime, root_url: str) -> Dict:
+def create_root_url_map(date: datetime, root_url: str) -> Dict:
     """Determine number of available pages for
     `date` YYYY-MM-DD represented by`root_url`.
 
     Args:
-        http (requests.Session): requests.Session with
-        connection timeout + connection retries with backoff
         date (datetime): Date for given `root_url`
         root_url (str): Root URL representing `date`
 
     Returns:
         Dict: Mapping of root URL to its total number of pages and its date
     """
+    # pylint: disable=broad-except
     root_url_to_last_page_and_date = dict()
-
+    first_page_url = root_url + "1"
     # Go to page 1
     try:
-        first_page_response = http.get(root_url + "1")
+        http = EnhancedSession().get_session()
+        first_page_response = http.get(first_page_url)
 
         soup = BeautifulSoup(first_page_response.content, "html.parser")
         # Find all instances of "/domains-registered-by-date/YYYY-MM-DD/{page_number}"
@@ -128,8 +68,8 @@ def create_root_url_map(http: requests.Session, date: datetime, root_url: str) -
             "num_pages": last_page,
             "date": date,
         }
-    except requests.exceptions.RequestException as error:
-        logger.error(error)
+    except Exception as error:
+        logger.error("%s %s", first_page_url, error)
     return root_url_to_last_page_and_date
 
 
@@ -155,7 +95,6 @@ def get_page_urls_by_date_str(root_urls_to_last_page_and_date: Dict) -> Dict:
 
 
 def download_domains(
-    http: requests.Session,
     date_str: str,
     page_urls: List[str],
     dataset_folder: str = "cubdomain_dataset",
@@ -164,13 +103,12 @@ def download_domains(
     for each day.
 
     Args:
-        http (requests.Session): requests.Session with
-        connection timeout + connection retries with backoff
         date_str (str): Date when domains were registered
         page_urls (List[str]): Page URLs containing domains registered on date `date_str`
         dataset_folder (str, optional): Folder to store domain .txt file in.
         Defaults to "cubdomain_dataset".
     """
+    # pylint: disable=broad-except
     if not os.path.exists(dataset_folder):
         os.mkdir(dataset_folder)
 
@@ -178,14 +116,15 @@ def download_domains(
         urls = set()
         for page_url in page_urls:
             try:
+                http = EnhancedSession().get_session()
                 page_response = http.get(page_url)
                 soup = BeautifulSoup(page_response.content, "html.parser")
                 # Each listed domain is encapsulated in this
                 # tag '<a href="https://www.cubdomain.com/site/ ...'
                 res = soup.find_all("a", href=lambda x: "cubdomain.com/site/" in x)
                 urls.update([line.string for line in res])
-            except requests.exceptions.RequestException as error:
-                logger.error(error)
+            except Exception as error:
+                logger.error("%s %s", page_url, error)
         file.write("\n".join(urls))
 
 
@@ -194,15 +133,7 @@ def scrape_cubdomain():
 
     dates, root_urls = generate_dates_and_root_urls()
     root_urls_to_last_page_and_date = dict(
-        ChainMap(
-            *execute_with_ray(
-                create_root_url_map,
-                [
-                    (EnhancedSession().get_session(), date, root_url)
-                    for date, root_url in zip(dates, root_urls)
-                ],
-            )
-        )
+        ChainMap(*execute_with_ray(create_root_url_map, list(zip(dates, root_urls))))
     )
 
     page_urls_by_date_str = get_page_urls_by_date_str(root_urls_to_last_page_and_date)
@@ -210,11 +141,7 @@ def scrape_cubdomain():
     execute_with_ray(
         download_domains,
         [
-            (EnhancedSession().get_session(), date_str, page_urls, "cubdomain_dataset")
+            (date_str, page_urls, "cubdomain_dataset")
             for date_str, page_urls in page_urls_by_date_str.items()
         ],
     )
-
-
-if __name__ == "__main__":
-    scrape_cubdomain()
